+++ src/Import/Eln.php
@@ -7,7 +7,7 @@
  * @package elabftw
  */
 
-namespace Elabftw\Services;
+namespace Elabftw\Import;
 
 use function basename;
 use Elabftw\Elabftw\CreateUpload;
@@ -26,7 +26,7 @@ use function sprintf;
 /**
  * Import a .eln file.
  */
-class ImportEln extends AbstractImportZip
+class Eln extends AbstractZip
 {
     // path where the metadata.json file lives (first folder found in archive)
     private string $root;
